% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kayaid.R
\name{kaya_identity}
\alias{kaya_identity}
\title{The Kaya Identity is an equation that expresses yearly CO2 emissions as a product of four factors}
\usage{
kaya_identity(pop, gdp, enInt, carbInt, output_type = "CO2")
}
\arguments{
\item{pop}{Population size (in millions)}

\item{gdp}{GDP per capita (in 1000$/person)}

\item{carbInt}{Carbon Intensity (in tonnes CO2/Gigajoule). Carbon Intensity is the CO2 emitted for produced energy. This number depends on the energy mix used (coal, solar, . . . ).}

\item{enINT}{Energy Intensity (in Gigajoule/$1000GDP). Energy Intensity is the energy needed to produce a certain amount of economic value.}
}
\value{
million tonnes of CO2 emissions
}
\description{
The Kaya Identity can be computed as: co2 = pop  gdp  enInt  carbInt. You can see by the units that they automatically resolve to million tonnes of CO2.
}
